requirements.txt
python-dotenv~=1.0.1
anthropic~=0.32.0
ipython~=8.26.0
pillow~=10.4.0
fastapi[all]~=0.112.0
uvicorn[standard]~=0.30.5
setuptools
python-multipart
redis~=5.0.8
opencv-python~=4.10.0.84

SQLAlchemy~=2.0.35
httpx~=0.27.0
Scrapy~=2.11.2
numpy~=2.0.1
pydantic~=2.8.2
passlib~=1.7.4
alembic~=1.13.2
python-jose
bcrypt==4.2.0
nats-py~=2.9.0
jinja2~=3.1.4

psycopg2
uvicorn

loguru~=0.7.2
pytest~=8.3.3
asyncpg

transformers~=4.45.2
psutil~=6.1.0

pydantic-settings
